add new header for npc_opinion and fix includes
