+++ src/npc_class.cpp
@@ -18,6 +18,7 @@
 #include "itype.h"
 #include "json.h"
 #include "mutation.h"
+#include "npc.h"
 #include "rng.h"
 #include "skill.h"
 #include "trait_group.h"
