+++ src/shop_cons_rate.cpp
@@ -8,6 +8,7 @@
 #include "item_group.h"
 #include "itype.h"
 #include "json.h"
+#include "npc.h"
 
 namespace
 {
